+++ test/js/components/tasks/capture/index.test.jsx
@@ -1,0 +1,251 @@
+import { fireEvent, waitForElement } from '@testing-library/react';
+import React from 'react';
+import { MemoryRouter } from 'react-router-dom';
+
+import CaptureModal from '@/components/tasks/capture';
+import { createObject } from '@/store/actions';
+
+import { renderWithRedux, storeWithThunk } from '../../../utils';
+
+jest.mock('@/store/actions');
+
+createObject.mockReturnValue(() =>
+  Promise.resolve({ type: 'TEST', payload: {} }),
+);
+
+afterEach(() => {
+  createObject.mockClear();
+});
+
+const defaultChangeset = {
+  Foo: ['Bar'],
+  Buz: ['Baz', 'Bing'],
+};
+
+describe('<CaptureModal/>', () => {
+  const setup = (options) => {
+    const defaults = {
+      changeset: defaultChangeset,
+    };
+    const opts = Object.assign({}, defaults, options);
+    const { changeset } = opts;
+    const toggleModal = jest.fn();
+    const result = renderWithRedux(
+      <MemoryRouter>
+        <CaptureModal
+          orgId="org-id"
+          changeset={changeset}
+          directories={{ config: ['foo/bar'] }}
+          isOpen
+          toggleModal={toggleModal}
+        />
+      </MemoryRouter>,
+      {},
+      storeWithThunk,
+    );
+    return { ...result, toggleModal };
+  };
+
+  test('can navigate forward/back, close modal', () => {
+    const { getByText, toggleModal } = setup();
+
+    expect(getByText('Select the location to capture changes')).toBeVisible();
+
+    fireEvent.click(getByText('Save & Next'));
+
+    expect(getByText('Select the changes to capture')).toBeVisible();
+
+    fireEvent.click(getByText('Go Back'));
+
+    expect(getByText('Select the location to capture changes')).toBeVisible();
+
+    fireEvent.click(getByText('Close'));
+
+    expect(toggleModal).toHaveBeenCalledTimes(1);
+  });
+
+  describe('form submit', () => {
+    test('creates a new commit', () => {
+      const { getByText, getByLabelText } = setup();
+
+      // Click forward to the select-changes modal:
+      fireEvent.click(getByText('Save & Next'));
+
+      const selectAll = getByLabelText('Select All Changes');
+      fireEvent.click(selectAll);
+      // Click forward to the commit-message modal:
+      fireEvent.click(getByText('Save & Next'));
+
+      const commitInput = getByLabelText('*Commit Message', { exact: false });
+      const submit = getByText('Capture Selected Changes');
+      fireEvent.change(commitInput, { target: { value: 'My Commit' } });
+      fireEvent.click(submit);
+
+      expect(getByText('Capturing Selected Changes…')).toBeVisible();
+      expect(createObject).toHaveBeenCalledTimes(1);
+      expect(createObject).toHaveBeenCalledWith({
+        objectType: 'scratch_org_commit',
+        url: window.api_urls.scratch_org_commit('org-id'),
+        data: {
+          commit_message: 'My Commit',
+          changes: defaultChangeset,
+          target_directory: 'src',
+        },
+        hasForm: true,
+        shouldSubscribeToObject: false,
+      });
+    });
+  });
+
+  describe('form error', () => {
+    test.each([
+      ['target_directory', 'Select the location to capture changes', true],
+      ['changes', 'Select the changes to capture', true],
+      ['commit_message', 'Describe the changes you are capturing', true],
+      ['foobar', 'Describe the changes you are capturing', false],
+    ])(
+      'navigates to correct page to show error: %s',
+      async (field, text, showsErr) => {
+        createObject.mockReturnValueOnce(() =>
+          // eslint-disable-next-line prefer-promise-reject-errors
+          Promise.reject({
+            body: {
+              [field]: ['Do not do that'],
+            },
+            response: {
+              status: 422,
+            },
+          }),
+        );
+        const { getByText, getByLabelText } = setup();
+        // Click forward to the select-changes modal:
+        fireEvent.click(getByText('Save & Next'));
+        const selectAll = getByLabelText('Select All Changes');
+        fireEvent.click(selectAll);
+        // Click forward to the commit-message modal:
+        fireEvent.click(getByText('Save & Next'));
+        const commitInput = getByLabelText('*Commit Message', { exact: false });
+        const submit = getByText('Capture Selected Changes');
+        fireEvent.change(commitInput, { target: { value: 'My Commit' } });
+        fireEvent.click(submit);
+
+        expect.assertions(showsErr ? 2 : 1);
+        await waitForElement(() => getByText(text));
+
+        expect(getByText(text)).toBeVisible();
+        if (showsErr) {
+          expect(getByText('Do not do that')).toBeVisible();
+        }
+      },
+    );
+  });
+
+  describe('select changes', () => {
+    let getters, selectAll, group1, inputs;
+
+    beforeEach(() => {
+      getters = setup();
+      const { getByLabelText, getByText } = getters;
+      fireEvent.click(getByText('Save & Next'));
+      selectAll = getByLabelText('Select All Changes');
+      group1 = getByLabelText('Buz');
+      inputs = [
+        getByLabelText('Bar'),
+        getByLabelText('Baz'),
+        getByLabelText('Bing'),
+      ];
+    });
+
+    describe('select-all/none', () => {
+      test('selects/deselects all items', () => {
+        fireEvent.click(selectAll);
+
+        expect(inputs[0].checked).toBe(true);
+        expect(inputs[1].checked).toBe(true);
+        expect(inputs[2].checked).toBe(true);
+
+        fireEvent.click(selectAll);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(false);
+        expect(inputs[2].checked).toBe(false);
+      });
+    });
+
+    describe('select/deselect-group', () => {
+      test('selects/deselects all items in group', () => {
+        fireEvent.click(group1);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(true);
+        expect(inputs[2].checked).toBe(true);
+
+        fireEvent.click(group1);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(false);
+        expect(inputs[2].checked).toBe(false);
+      });
+    });
+
+    describe('select/deselect-item', () => {
+      test('selects/deselects item', () => {
+        fireEvent.click(inputs[1]);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(true);
+        expect(inputs[2].checked).toBe(false);
+
+        fireEvent.click(inputs[2]);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(true);
+        expect(inputs[2].checked).toBe(true);
+
+        fireEvent.click(inputs[2]);
+
+        expect(inputs[0].checked).toBe(false);
+        expect(inputs[1].checked).toBe(true);
+        expect(inputs[2].checked).toBe(false);
+      });
+    });
+
+    describe('accordion panel click', () => {
+      test('expands/collapses', () => {
+        const { baseElement, getByTitle } = getters;
+        const content = baseElement.querySelector('.slds-accordion__content');
+
+        expect(content).toHaveAttribute('aria-hidden', 'true');
+
+        fireEvent.click(getByTitle('Buz'));
+
+        expect(content).toHaveAttribute('aria-hidden', 'false');
+      });
+    });
+  });
+
+  describe('commit message', () => {
+    let getters;
+
+    beforeEach(() => {
+      getters = setup();
+      const { getByLabelText, getByText } = getters;
+      fireEvent.click(getByText('Save & Next'));
+      fireEvent.click(getByLabelText('Select All Changes'));
+      fireEvent.click(getByText('Save & Next'));
+    });
+
+    describe('accordion panel click', () => {
+      test('expands/collapses', () => {
+        const { baseElement, getByTitle } = getters;
+        const content = baseElement.querySelector('.slds-accordion__content');
+
+        expect(content).toHaveAttribute('aria-hidden', 'true');
+
+        fireEvent.click(getByTitle('Buz'));
+
+        expect(content).toHaveAttribute('aria-hidden', 'false');
+      });
+    });
+  });
+});
